Add missing properties to test classes
Ported from https://github.com/Shardj/zf1-future/pull/289

PHP 8.2: Dynamic Properties are deprecated fix

Reference: https://php.watch/versions/8.2/dynamic-properties-deprecated

+++ tests/Zend/Mail/MboxMessageOldTest.php
@@ -93,7 +93,9 @@ class Zend_Mail_MboxMessageOldTest extends PHPUnit_Framework_TestCase
 
     public function tearDown()
     {
-        unlink($this->_mboxFile);
+        if ($this->_mboxFile) {
+            unlink($this->_mboxFile);
+        }
     }
 
     public function testFetchHeader()
